+++ changelog.md
@@ -3,6 +3,7 @@
 ## Next version
 - Refactoring :
   - [PRE-1983](https://git.payplug.com/plugins/prestashop_v2_1.7/-/merge_requests/2279): Add action Payment
+  - [PRE-1983](https://git.payplug.com/plugins/prestashop_v2_1.7/-/merge_requests/2279): Add action Payment::abort && Payment::capture
 
 ## Version 4.5.0